requirements.txt
aiogram>=3.0.0
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
httpx
python-dotenv
